refactor(container/components): update container imports
