+++ client/src/components/Lesson/Instructions/Instructions.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 import { AppState } from '../../../store/configureStore';
