+++ client/src/components/Login/FormTemplate/FormTemplate.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import Login from '../../../containers/Login/Login';
-import { useLocation } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
+import { Login } from '../../../containers';
 
 export default function FormTemplate(props: any): JSX.Element {
   const path: string = useLocation().pathname;
